backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
flask-cors==3.0.10
python-whois==0.7.3
requests==2.26.0
